+++ src/my_first_adventure_game/game/entities/__init__.py
@@ -1,5 +1,6 @@
 from my_first_adventure_game.game.entities.enemy import Enemy
 from my_first_adventure_game.game.entities.npc import NPC
+from my_first_adventure_game.game.entities.npc_movement import move_npc_towards
 from my_first_adventure_game.game.entities.player import Player
 
-__all__ = ["Enemy", "NPC", "Player"]
+__all__ = ["Enemy", "NPC", "Player", "move_npc_towards"]